Island: Move UserCreds to models package
