+++ monkey/monkey_island/cc/models/__init__.py
@@ -9,3 +9,4 @@ from .pba_results import PbaResults
 from monkey_island.cc.models.report.report import Report
 from .stolen_credentials import StolenCredentials
 from .simulation import Simulation, SimulationSchema, IslandMode
+from .user_credentials import UserCreds
