+++ monkey/monkey_island/cc/models/user_credentials.py
