+++ monkey/monkey_island/cc/repository/i_user_repository.py
@@ -1,6 +1,6 @@
 import abc
 
-from monkey_island.cc.services.authentication.user_creds import UserCreds
+from monkey_island.cc.models import UserCreds
 
 
 class IUserRepository(metaclass=abc.ABCMeta):
