+++ monkey/monkey_island/cc/services/authentication/json_file_user_datastore.py
@@ -6,10 +6,9 @@ from common.utils.exceptions import (
     InvalidRegistrationCredentialsError,
     UnknownUserError,
 )
+from monkey_island.cc.models import UserCreds
 from monkey_island.cc.repository import IUserRepository
 from monkey_island.cc.server_utils.file_utils import open_new_securely_permissioned_file
-
-from .user_creds import UserCreds
 
 CREDENTIALS_FILE = "credentials.json"
